Add clients, vaults configuration to pillar
Enable use of pillar in the state files that depend
directly on the clients and vaults configuration.

Note: the configuration of the SSH socket (SSH_VAULT_VM) currently
limits the number of vault qubes to one. All the clients will be
configured to use the vault that is defined first.

The configuration file is identified as such and should be kept
when the package is reinstalled.