Work around Salt pillar limitation in Qubes OS
Despite enabling /srv/user_pillar as pillar_root
in /etc/salt/master.d/user.conf, files are not found
when included from a directory in /srv/user_pillar.

Creating symbolic links in /srv/pillar works around
the issue.

The symlink to /srv/user_pillar/split-ssh allows files
to be included form Jinja templates, and the symlink
names /srv/pillar/_tops/split-ssh.top ensures that
the pillar is enabled.

Since this is not a formula limitation, I think it is
appropriate to work around the issue in the package spec
and not the formula Makefile.